+++ GeneticAlgorithmBase/ChromoKnapsack.cpp
@@ -114,7 +114,7 @@ namespace ga
 		encodedSize += sizeof(bool) * hasItem.size();
 
 		// Reserve string memory
-		m_encoded.reserve(encodedSize + static_cast<std::size_t>(1));
+		m_encoded.reserve(encodedSize + 1);
 
 		m_encoded += encodeBoolVector(hasItem);
 	}
